Last Commit
Cannot spend any more time on this as I have a report to write. This is
the edition that will used for the viva (unless I find more time to
spend coding and potentially fix some stuff)

Only 3 features aren't as I want them within this engine
1) Shadows, the shadow code does not work but as it is openGL it is very
difficult to debug and I haven't got the time to do that
2) Collisions, the collision system I have is simple and does not
properly calculate penetration depth and doesn't calculate collision
normals
3) Because of the collisions the physics system is not as advanced as I
want it to be because I cannot do impulse resolution with the way the
collisions are.

If I had another month to code I could finish all these elements up and
also make more gameplay to demonstrate these features